refactor(view): split design_import.hpp into focused headers behind umbrella
Zero-behavior-change packaging refactor (PR-2 of
planning/2026-05-29-frontend-ir-refactor-plan.md). The 1097-line public
umbrella design_import.hpp is split into design_ir.hpp (IR model/types),
design_tokens.hpp, design_sources.hpp, design_shortcuts.hpp, and
design_codegen.hpp, with design_import.hpp reduced to a 159-line umbrella that
re-includes them in dependency order (acyclic: design_ir is foundational;
tokens/sources/shortcuts/codegen include it; codegen also includes shortcuts).
DesignSource lives in design_ir.hpp (DesignIR::source needs the full enum). The
native-binding section (widget forward decls, NativeImport* descriptors,
build_native_view_tree) is retained verbatim in the umbrella for now. All 84
includers compile unchanged; every top-level decl preserved (verified by diff);
all 18 design-import test binaries pass.

Skill-Update: skip skill=import-design reason="header packaging split behind a back-compat umbrella; no behavior or API change"
